Update check-aggregate.rb to support honor-stash
The honor-stash function needs a list of maps with check/client/status information. This change includes the queries to add that information to when calling the named_aggregate_results function.

I'm not a very experienced ruby programmer so feel free to reject and treat this as a bug report.

Thanks for the plugins.